[이중보] - 이전 문제였던 제품 상세 조회 api null값과 예외사항 발생사항 수정완성 - isbn13타입을 long으로 고쳐완성
#6

+++ src/main/java/com/jung/Ssave/ssave/domain/Item.java
@@ -15,7 +15,7 @@ public class Item {
 
 	private long itemId;
 	
-	private int isbn13;
+	private long isbn13;
 	
 	private String link;
